backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.122.0",
    "gunicorn>=23.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
    "email-validator>=2.3.0",
]
